Fixed a problem with Russian Roulette and paths with specular vertices

+++ src/slg/engines/pathcpu/pathcpu.cpp
@@ -40,8 +40,8 @@ void PathCPURenderEngine::StartLockLess() {
 	// Rendering parameters
 	//--------------------------------------------------------------------------
 
-	maxPathDepth = Max(1, cfg.Get(Property("path.maxdepth")(5)).Get<int>());
-	rrDepth = Max(1, cfg.Get(Property("path.russianroulette.depth")(3)).Get<int>());
+	maxPathDepth = (u_int)Max(1, cfg.Get(Property("path.maxdepth")(5)).Get<int>());
+	rrDepth = (u_int)Max(1, cfg.Get(Property("path.russianroulette.depth")(3)).Get<int>());
 	rrImportanceCap = Clamp(cfg.Get(Property("path.russianroulette.cap")(.5f)).Get<float>(), 0.f, 1.f);
 
 	// Clamping settings
